+++ wirestead/builder/tcp_client_builder.cc
@@ -25,8 +25,7 @@
 namespace wirestead {
 namespace builder {
 
-template <uint32_t State>
-TcpClientBuilder<State>::TcpClientBuilder(const std::string& host, uint16_t port)
+TcpClientBuilder::TcpClientBuilder(const std::string& host, uint16_t port)
     : host_(host),
       port_(port),
       auto_start_(false),
@@ -48,8 +47,7 @@ TcpClientBuilder<State>::TcpClientBuilder(const std::string& host, uint16_t port
   AutoInitializer::ensure_io_context_running();
 }
 
-template <uint32_t State>
-std::unique_ptr<wrapper::TcpClient> TcpClientBuilder<State>::build() {
+std::unique_ptr<wrapper::TcpClient> TcpClientBuilder::build() {
   std::unique_ptr<wrapper::TcpClient> client;
   if (independent_context_) {
     client = std::make_unique<wrapper::TcpClient>(host_, port_, std::make_shared<boost::asio::io_context>());
@@ -96,93 +94,77 @@ std::unique_ptr<wrapper::TcpClient> TcpClientBuilder<State>::build() {
   return client;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::auto_start(bool auto_start) {
+TcpClientBuilder& TcpClientBuilder::auto_start(bool auto_start) {
   auto_start_ = auto_start;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::retry_interval(std::chrono::milliseconds interval) {
+TcpClientBuilder& TcpClientBuilder::retry_interval(std::chrono::milliseconds interval) {
   retry_interval_ = interval;
   retry_interval_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::max_retries(int max_retries) {
+TcpClientBuilder& TcpClientBuilder::max_retries(int max_retries) {
   max_retries_ = max_retries;
   max_retries_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::connection_timeout(std::chrono::milliseconds timeout) {
+TcpClientBuilder& TcpClientBuilder::connection_timeout(std::chrono::milliseconds timeout) {
   connection_timeout_ = timeout;
   connection_timeout_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::idle_timeout(std::chrono::milliseconds timeout) {
+TcpClientBuilder& TcpClientBuilder::idle_timeout(std::chrono::milliseconds timeout) {
   idle_timeout_ = timeout;
   idle_timeout_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::idle_timeout_action(IdleTimeoutAction action) {
+TcpClientBuilder& TcpClientBuilder::idle_timeout_action(IdleTimeoutAction action) {
   idle_timeout_action_ = action;
   idle_timeout_action_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::independent_context(bool use_independent) {
+TcpClientBuilder& TcpClientBuilder::independent_context(bool use_independent) {
   independent_context_ = use_independent;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::tcp_no_delay(bool enable) {
+TcpClientBuilder& TcpClientBuilder::tcp_no_delay(bool enable) {
   tcp_no_delay_ = enable;
   tcp_no_delay_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::keep_alive(bool enable) {
+TcpClientBuilder& TcpClientBuilder::keep_alive(bool enable) {
   keep_alive_ = enable;
   keep_alive_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::send_buffer_size(size_t bytes) {
+TcpClientBuilder& TcpClientBuilder::send_buffer_size(size_t bytes) {
   send_buffer_size_ = bytes;
   send_buffer_size_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::receive_buffer_size(size_t bytes) {
+TcpClientBuilder& TcpClientBuilder::receive_buffer_size(size_t bytes) {
   receive_buffer_size_ = bytes;
   receive_buffer_size_set_ = true;
   return *this;
 }
 
-template <uint32_t State>
-TcpClientBuilder<State>& TcpClientBuilder<State>::read_buffer_size(size_t bytes) {
+TcpClientBuilder& TcpClientBuilder::read_buffer_size(size_t bytes) {
   read_buffer_size_ = bytes;
   read_buffer_size_set_ = true;
   return *this;
 }
 
 // Explicit template instantiations
-template class TcpClientBuilder<BuilderState::None>;
-template class TcpClientBuilder<BuilderState::HasData>;
-template class TcpClientBuilder<BuilderState::HasError>;
-template class TcpClientBuilder<BuilderState::Ready>;
 
 }  // namespace builder
 }  // namespace wirestead