+++ wirestead/builder/tcp_client_builder.hpp
@@ -36,55 +36,9 @@ namespace builder {
 /**
  * @brief Modernized Builder for TcpClient
  */
-template <uint32_t State = BuilderState::None>
-class WIRESTEAD_API TcpClientBuilder : public BuilderInterface<wrapper::TcpClient, TcpClientBuilder<State>, State> {
+class WIRESTEAD_API TcpClientBuilder : public BuilderInterface<wrapper::TcpClient, TcpClientBuilder> {
  public:
-  template <uint32_t NewState>
-  using Rebind = TcpClientBuilder<NewState>;
-
   TcpClientBuilder(const std::string& host, uint16_t port);
-
-  // Allow conversion between states
-  template <uint32_t OtherState>
-  TcpClientBuilder(TcpClientBuilder<OtherState>&& other) noexcept
-      : host_(std::move(other.host_)),
-        port_(other.port_),
-        auto_start_(other.auto_start_),
-        independent_context_(other.independent_context_),
-        retry_interval_(other.retry_interval_),
-        retry_interval_set_(other.retry_interval_set_),
-        max_retries_(other.max_retries_),
-        max_retries_set_(other.max_retries_set_),
-        connection_timeout_(other.connection_timeout_),
-        connection_timeout_set_(other.connection_timeout_set_),
-        idle_timeout_(other.idle_timeout_),
-        idle_timeout_set_(other.idle_timeout_set_),
-        idle_timeout_action_(other.idle_timeout_action_),
-        idle_timeout_action_set_(other.idle_timeout_action_set_),
-        tcp_no_delay_(other.tcp_no_delay_),
-        tcp_no_delay_set_(other.tcp_no_delay_set_),
-        keep_alive_(other.keep_alive_),
-        keep_alive_set_(other.keep_alive_set_),
-        send_buffer_size_(other.send_buffer_size_),
-        send_buffer_size_set_(other.send_buffer_size_set_),
-        receive_buffer_size_(other.receive_buffer_size_),
-        receive_buffer_size_set_(other.receive_buffer_size_set_),
-        read_buffer_size_(other.read_buffer_size_),
-        read_buffer_size_set_(other.read_buffer_size_set_) {
-    this->on_data_ = std::move(other.on_data_);
-    this->on_error_ = std::move(other.on_error_);
-    this->on_connect_ = std::move(other.on_connect_);
-    this->on_disconnect_ = std::move(other.on_disconnect_);
-    this->on_message_ = std::move(other.on_message_);
-    this->on_data_batch_ = std::move(other.on_data_batch_);
-    this->on_message_batch_ = std::move(other.on_message_batch_);
-    this->on_backpressure_ = std::move(other.on_backpressure_);
-    this->framer_factory_ = std::move(other.framer_factory_);
-    this->bp_strategy_ = other.bp_strategy_;
-    this->bp_threshold_ = other.bp_threshold_;
-    this->bp_strategy_set_ = other.bp_strategy_set_;
-    this->bp_threshold_set_ = other.bp_threshold_set_;
-  }
 
   // Delete copy
   TcpClientBuilder(const TcpClientBuilder&) = delete;
@@ -92,30 +46,30 @@ class WIRESTEAD_API TcpClientBuilder : public BuilderInterface<wrapper::TcpClien
 
   std::unique_ptr<wrapper::TcpClient> build() override;
 
-  TcpClientBuilder<State>& auto_start(bool auto_start = true) override;
+  TcpClientBuilder& auto_start(bool auto_start = true) override;
 
-  TcpClientBuilder<State>& retry_interval(std::chrono::milliseconds interval);
-  TcpClientBuilder<State>& max_retries(int max_retries);
-  TcpClientBuilder<State>& connection_timeout(std::chrono::milliseconds timeout);
+  TcpClientBuilder& retry_interval(std::chrono::milliseconds interval);
+  TcpClientBuilder& max_retries(int max_retries);
+  TcpClientBuilder& connection_timeout(std::chrono::milliseconds timeout);
   /**
    * @brief Configure application-level idle timeout.
    *
    * A value of 0ms disables idle timeout. When enabled, inbound or outbound
    * activity resets the timer.
    */
-  TcpClientBuilder<State>& idle_timeout(std::chrono::milliseconds timeout);
+  TcpClientBuilder& idle_timeout(std::chrono::milliseconds timeout);
   /**
    * @brief Configure what happens when an enabled idle timeout expires.
    *
    * The default is IdleTimeoutAction::Reconnect. This setting has no effect
    * while idle_timeout is 0ms.
    */
-  TcpClientBuilder<State>& idle_timeout_action(IdleTimeoutAction action);
-  TcpClientBuilder<State>& independent_context(bool use_independent = true);
-  TcpClientBuilder<State>& tcp_no_delay(bool enable = true);
-  TcpClientBuilder<State>& keep_alive(bool enable = true);
-  TcpClientBuilder<State>& send_buffer_size(size_t bytes);
-  TcpClientBuilder<State>& receive_buffer_size(size_t bytes);
+  TcpClientBuilder& idle_timeout_action(IdleTimeoutAction action);
+  TcpClientBuilder& independent_context(bool use_independent = true);
+  TcpClientBuilder& tcp_no_delay(bool enable = true);
+  TcpClientBuilder& keep_alive(bool enable = true);
+  TcpClientBuilder& send_buffer_size(size_t bytes);
+  TcpClientBuilder& receive_buffer_size(size_t bytes);
 
   /**
    * @brief Size of the userspace buffer each read fills, in bytes.
@@ -124,12 +78,9 @@ class WIRESTEAD_API TcpClientBuilder : public BuilderInterface<wrapper::TcpClien
    * transfers, at the cost of that much memory per connection. Clamped to
    * [MIN_READ_BUFFER_SIZE, MAX_READ_BUFFER_SIZE].
    */
-  TcpClientBuilder<State>& read_buffer_size(size_t bytes);
+  TcpClientBuilder& read_buffer_size(size_t bytes);
 
  private:
-  template <uint32_t S>
-  friend class TcpClientBuilder;
-
   std::string host_;
   uint16_t port_;
   bool auto_start_;
@@ -157,7 +108,7 @@ class WIRESTEAD_API TcpClientBuilder : public BuilderInterface<wrapper::TcpClien
   bool read_buffer_size_set_{false};
 };
 
-using TcpClientBuilderDefault = TcpClientBuilder<BuilderState::None>;
+using TcpClientBuilderDefault = TcpClientBuilder;
 
 #ifdef _MSC_VER
 #pragma warning(pop)